invalidate cache

+++ lib/cache.ts
@@ -24,9 +24,19 @@ function getFilePath(name: string, network: string): string {
 }
 
 export const cacheDel = (path: string): void => {
-  logger.debug('Delete cache', path);
-  fs.unlinkSync(path);
-  memoryCache[path] = null;
+  const isRedisCache = !!config.REDIS_URL;
+
+  if (isRedisCache && redis) {
+    redis?.del(path);
+  } else {
+    try {
+      logger.debug('cacheDel: ', path);
+      memoryCache[path] = null;
+      fs.unlinkSync(path);
+    } catch (e) {
+      logger.error(`cacheDel: ${e.message}`);
+    }
+  }
 };
 
 export const cacheGet = async (
